# backend/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Microsoft Engage Backend",
    "description": "This is Microsoft Engage Backend server.  You can find out more about     Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).      For this sample, you can use the api key `special-key` to test the authorization     filters.",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "archeelparekh11@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "https://microsoft-engage-2021.herokuapp.com"
    },
    {
      "url": "http://localhost:4000"
    }
  ],
...
  "paths": {
...
    "/teacher/course/{courseId}/assignment/{assignmentId}/submission/{submissionId}": {
      "post": {
        "tags": ["Teacher"],
        "summary": "Grade a submission for assignment",
        "description": "To get a submission for an assignment by the authorised teacher.",
        "operationId": "gradeSubmission",
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "description": "Course ID of the course created by teacher",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "assignmentId",
            "in": "path",
            "description": "assignment ID of the assignment created by teacher",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "submissionId",
            "in": "path",
            "description": "submission ID of the assignment submitted by teacher",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Grade",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "grade": {
                    "type": "integer"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Assignment created successfully",
            "content": {}
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-codegen-request-body-name": "body"
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}